# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Loginex TMS API
  description: |
    API для диспетчерской системы Loginex TMS.
    - Auth: серверные сессии в БД. В cookie — подписанный HMAC-SHA256 токен с jti = id строки Session
      (loginex_staff_session для сотрудников, loginex_driver_session для водителей).
      Отзыв строки Session немедленно закрывает доступ.
    - Роли: admin | logist (кабинет), driver (мобильное приложение).
      driverId и принадлежность организации берутся из сессии, а не из тела запроса.
    - Rate limiting: 5 попыток / 15 мин для auth
    - CSRF: double-submit cookie loginex_csrf + header x-csrf-token
    - Validation: Zod, 400 с details.issues
  version: 1.0.0
  contact:
    name: Loginex TMS
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://loginex.example.com
    description: Production
...
paths:
...
  /api/auth/users/{id}:
    patch:
      tags: [Admin]
      summary: Одобрить заявку / заблокировать / восстановить / сменить роль (admin)
      description: Действие пишется в AuditLog.
      parameters:
        - in: path
          name: id
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Пользователь обновлён
        '400':
          description: Некорректное действие
        '403':
          description: Недостаточно прав
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Staff login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: admin@loginex.ru
                password:
                  type: string
                  example: demo_dev_only
      responses:
        '200':
          description: Logged in, sets cookies
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid credentials
        '429':
          description: Rate limited
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user (pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
      responses:
        '200':
          description: Registered
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email exists
...
  /api/admin/audit:
    get:
      tags: [Admin]
      summary: Audit logs (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: action
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Logs
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: loginex_token
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, dispatcher, manager, viewer]
        status:
          type: string
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Validation failed
              details:
                type: object
                properties:
                  issues:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: array
                          items:
                            type: string
                        message:
                          type: string
...